# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Necter Hub API
  version: 1.0.0
  summary: Public JSON API of the Necter testnet (testnet-rpc.necter.network) and its internal admin surface.
  description: |
    Normative companion of `spec/PLATFORM.md`. Section references (§a … §k) point there.

    **Conventions**
    * Every response is `application/json; charset=utf-8` except the two documented binary downloads
      (`GET /v1/modules/{address}/hbc`, `GET /v1/assets/{sha256}`). Unknown paths → 404 JSON.
    * Field names are snake_case. Timestamps are unix seconds (`*_at`). Token amounts are decimal strings in
      base units (wei); token objects carry `decimals`. Hashes, ids and addresses are lowercase `0x` hex.
    * Lists use cursor pagination: `?limit=1..200` (default 50) and `?cursor=<opaque>`; responses are
      `{"items": [...], "next_cursor": "…" | null}`.
    * Errors: `{"error": "<code>", "message": "…", "details": {…}, "request_id": "…"}` with the HTTP status
      listed per endpoint. `error` codes are stable; `message` is for humans.
    * Rate limits: every response carries `RateLimit-Limit`, `RateLimit-Remaining`, `RateLimit-Reset`
      (seconds); 429 adds `Retry-After`. Budgets are per API key, per session wallet, per node key, and per
      IP for anonymous calls.
    * Auth:
      - `session`: bearer token from SIWE (EIP-4361) wallet sign-in (`/v1/auth/*`), 12 h lifetime.
...
        `X-Ndsr-Nonce`, `X-Ndsr-Signature`), signed over the path as sent.
...
  license:
...
servers:
...
security: []
paths:
...
  /v1/projects/{project_id}:
    get:
      tags: [projects]
      operationId: getProject
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: 'nsess_<43 char base64url>'
      description: SIWE session token (POST /v1/auth/siwe), 12 h, revocable.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: 'nk_test_<8>_<43>'
      description: Developer API key; scopes in ApiScope. Stored as SHA-256 on the Hub.
    nodeSig:
      type: apiKey
      in: header
      name: X-Ndsr-Signature
      description: PROTOCOL.md node request signing; also requires X-Ndsr-Node, X-Ndsr-Public-Key, X-Ndsr-Timestamp, X-Ndsr-Nonce.
    operator:
      type: http
      scheme: bearer
      description: Operator session on gateway.testnet.necter.network only.
...
  parameters:
...
    ProjectId: { name: project_id, in: path, required: true, schema: { $ref: '#/components/schemas/Hash' } }
...
  responses:
...
    NotFound: { description: '`not_found`', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
...
    Address: { type: string, pattern: '^0x[0-9a-f]{40}$', description: Lowercase EVM address }
    Hash: { type: string, pattern: '^0x[0-9a-f]{64}$' }
    Amount: { type: string, pattern: '^(0|[1-9][0-9]{0,77})$', description: Token base units }
    UnixTime: { type: integer, minimum: 0 }
...
    Category: { type: string, enum: [blockchain, depin, machine-learning, storage, compute, bandwidth, data-sovereignty, iot, hardware-staking, content-delivery] }
    DeviceClass: { type: string, enum: [phone, tablet, laptop, desktop, server] }
...
    Engine: { type: string, enum: [native, pulley64, pulley32] }
...
    ListingStatus: { type: string, enum: [draft, submitted, pending_review, listed, paused, underfunded, rejected, delisted] }
...
    Error:
      type: object
      required: [error, message, request_id]
      properties:
        error: { type: string, description: Stable machine code }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        request_id: { type: string }

    # ── network ──
...
    Token:
      type: object
      required: [address, symbol, decimals]
      properties:
        address: { $ref: '#/components/schemas/Address' }
        symbol: { type: string }
        name: { type: string }
        decimals: { type: integer, minimum: 0, maximum: 36 }
...
    ProjectSummary:
      type: object
      required: [project_id, slug, developer, name, category, icon, listing_status, token]
      properties:
        project_id: { $ref: '#/components/schemas/Hash' }
        slug: { type: string }
        developer: { $ref: '#/components/schemas/Address' }
        developer_name: { type: [string, 'null'] }
        developer_verified: { type: boolean }
        name: { type: string }
        tagline: { type: [string, 'null'] }
        category: { $ref: '#/components/schemas/Category' }
        tags: { type: array, items: { type: string } }
        icon: { type: string, format: uri }
        accent_color: { type: [string, 'null'] }
        listing_status: { $ref: '#/components/schemas/ListingStatus' }
        device_classes: { type: array, items: { $ref: '#/components/schemas/DeviceClass' } }
        token: { $ref: '#/components/schemas/Token' }
        reward_model: { type: string, enum: [per-unit, epoch-pool] }
        min_collateral: { $ref: '#/components/schemas/Amount' }
        miners: { type: integer }
        avg_daily_reward_per_miner: { $ref: '#/components/schemas/Amount' }
        average_rating_x100: { type: [integer, 'null'] }
        review_count: { type: integer }
        trending: { type: boolean }
        featured: { type: boolean }
        listed_at: { oneOf: [ { $ref: '#/components/schemas/UnixTime' }, { type: 'null' } ] }
    Project:
      allOf:
        - { $ref: '#/components/schemas/ProjectSummary' }
        - type: object
          required: [version, manifest_hash, consensus_hash, consensus, scheduling, listing]
          properties:
            version: { type: integer }
            manifest_hash: { $ref: '#/components/schemas/Hash' }
            consensus_hash: { $ref: '#/components/schemas/Hash' }
            pending_version: { oneOf: [ { $ref: '#/components/schemas/ProjectVersion' }, { type: 'null' } ] }
            consensus: { $ref: '#/components/schemas/ManifestConsensus' }
            scheduling: { $ref: '#/components/schemas/ManifestScheduling' }
            listing: { $ref: '#/components/schemas/ManifestListing' }
            vault: { $ref: '#/components/schemas/Address' }
            registry_tx: { $ref: '#/components/schemas/Hash' }
            review: { type: object, properties: { status: { type: string }, reason: { type: [string, 'null'] }, conditions: { type: array, items: { type: string } }, decided_at: { type: [integer, 'null'] } } }
            stats: { $ref: '#/components/schemas/ProjectStats' }
            economics: { $ref: '#/components/schemas/ProjectEconomics' }
            created_at: { $ref: '#/components/schemas/UnixTime' }
...
    ManifestConsensus:
      type: object
      additionalProperties: false
      required: [modules, work, economics]
      properties:
        modules:
          type: object
          additionalProperties: false
          required: [worker, verifier]
          properties:
            worker: { $ref: '#/components/schemas/Hash' }
            verifier: { type: 'null' }
        work:
          type: object
          additionalProperties: false
          required: [class, verification, execution, functions, task_source, max_gas_limit, committee, round_secs, lease_secs, epoch_secs, dispute_window_secs]
          properties:
            class: { type: string, const: deterministic }
            verification: { type: string, const: redundant-execution }
            execution: { type: string, const: committee }
            functions: { type: array, minItems: 1, maxItems: 64, items: { type: string } }
            task_source:
              oneOf:
                - { type: object, additionalProperties: false, required: [kind], properties: { kind: { const: api } } }
                - { type: object, additionalProperties: false, required: [kind, function, interval_secs, gas_limit], properties: { kind: { const: schedule }, function: { type: string }, interval_secs: { type: integer, minimum: 10, maximum: 3600 }, gas_limit: { type: integer, minimum: 1 } } }
            max_gas_limit: { type: integer, minimum: 1, maximum: 10000000000 }
            committee:
              type: object
              additionalProperties: false
              required: [size, backups, min_size, selection, weight_cap_multiple]
              properties:
                size: { type: integer, minimum: 3, maximum: 31 }
                backups: { type: integer, minimum: 0, maximum: 16 }
                min_size: { type: integer, minimum: 3, maximum: 31 }
                selection: { type: string, const: collateral-reputation-sortition-v1 }
                weight_cap_multiple: { type: integer, minimum: 1, maximum: 100 }
            round_secs: { type: integer, minimum: 10, maximum: 600 }
            lease_secs: { type: integer, minimum: 5, maximum: 600 }
            epoch_secs: { type: integer, enum: [3600, 7200, 10800, 14400, 21600, 28800, 43200, 86400] }
            dispute_window_secs: { type: integer, minimum: 3600, maximum: 604800 }
        economics:
          type: object
          additionalProperties: false
          required: [reward_token, vault, reward_model, reward_per_unit, daily_emission, fee_split_bp, max_units_per_epoch, max_recipients_per_epoch, min_collateral, slashing_bp]
          properties:
            reward_token: { $ref: '#/components/schemas/Address' }
            vault: { $ref: '#/components/schemas/Address' }
            reward_model: { type: string, enum: [per-unit, epoch-pool] }
            reward_per_unit: { $ref: '#/components/schemas/Amount' }
            daily_emission: { $ref: '#/components/schemas/Amount' }
            fee_split_bp:
              type: object
              additionalProperties: false
              required: [miner, developer, treasury]
              properties:
                miner: { type: integer, minimum: 5000, maximum: 10000 }
                developer: { type: integer, minimum: 0, maximum: 5000 }
                treasury: { type: integer, minimum: 0, maximum: 5000 }
            max_units_per_epoch: { type: integer, minimum: 1, maximum: 1000000000 }
            max_recipients_per_epoch: { type: integer, minimum: 1, maximum: 10000 }
            min_collateral: { $ref: '#/components/schemas/Amount' }
            slashing_bp:
              type: object
              additionalProperties: false
              required: [invalid_result, missed_sla]
              properties:
                invalid_result: { type: integer, minimum: 0, maximum: 5000 }
                missed_sla: { type: integer, minimum: 0, maximum: 500 }
    ManifestScheduling:
      type: object
      additionalProperties: false
      required: [requirements, sla, public_tasks]
      properties:
        requirements:
          type: object
          additionalProperties: false
          required: [device_classes, cpu_cores, ram_mb, storage_mb, gpu, min_benchmark, engines, attestation]
          properties:
            device_classes: { type: array, minItems: 1, items: { $ref: '#/components/schemas/DeviceClass' } }
            cpu_cores: { type: integer, minimum: 0 }
            ram_mb: { type: integer, minimum: 0 }
            storage_mb: { type: integer, minimum: 0 }
            gpu: { type: 'null' }
            min_benchmark: { type: integer, minimum: 0 }
            engines: { type: array, minItems: 1, items: { $ref: '#/components/schemas/Engine' } }
            attestation: { type: array, maxItems: 0 }
        sla:
          type: object
          additionalProperties: false
          required: [max_latency_ms, min_uptime_bp]
          properties:
            max_latency_ms: { type: integer, minimum: 100, maximum: 600000 }
            min_uptime_bp: { type: integer, minimum: 0, maximum: 10000 }
        public_tasks: { type: boolean }
    ManifestListing:
      type: object
      additionalProperties: false
      required: [name, tagline, description, category, tags, icon, banner, screenshots, video, website, docs, support, features, accent_color]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        tagline: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 8000 }
        category: { $ref: '#/components/schemas/Category' }
        tags: { type: array, maxItems: 8, items: { type: string, pattern: '^[a-z0-9-]{1,24}$' } }
        icon: { type: string, format: uri }
        banner: { type: [string, 'null'], format: uri }
        screenshots: { type: array, maxItems: 8, items: { type: string, format: uri } }
        video: { type: [string, 'null'], format: uri }
        website: { type: [string, 'null'], format: uri }
        docs: { type: [string, 'null'], format: uri }
        support: { type: [string, 'null'], format: uri }
        features: { type: array, maxItems: 12, items: { type: string, maxLength: 120 } }
        accent_color: { type: [string, 'null'], pattern: '^#[0-9A-F]{6}$' }
...
    ProjectVersion:
      type: object
      required: [version, manifest_hash, consensus_hash, tiers_changed, status, submitted_at]
      properties:
        version: { type: integer }
        manifest_hash: { $ref: '#/components/schemas/Hash' }
        consensus_hash: { $ref: '#/components/schemas/Hash' }
        tiers_changed: { type: array, items: { type: string, enum: [consensus, scheduling, listing] } }
        status: { type: string, enum: [submitted, registered, pending, active, superseded, rejected] }
        publish_tx: { oneOf: [ { $ref: '#/components/schemas/Hash' }, { type: 'null' } ] }
        published_at: { oneOf: [ { $ref: '#/components/schemas/UnixTime' }, { type: 'null' } ] }
        effective_epoch: { type: [integer, 'null'] }
        submitted_at: { $ref: '#/components/schemas/UnixTime' }
    ProjectEconomics:
      type: object
      required: [token, reward_model, reward_per_unit, daily_emission, epoch_cap, fee_split_bp, vault]
      properties:
        token: { $ref: '#/components/schemas/Token' }
        reward_model: { type: string, enum: [per-unit, epoch-pool] }
        reward_per_unit: { $ref: '#/components/schemas/Amount' }
        daily_emission: { $ref: '#/components/schemas/Amount' }
        epoch_secs: { type: integer }
        epoch_cap: { $ref: '#/components/schemas/Amount' }
        fee_split_bp: { type: object, properties: { miner: { type: integer }, developer: { type: integer }, treasury: { type: integer } } }
        min_collateral: { $ref: '#/components/schemas/Amount' }
        slashing_bp: { type: object, properties: { invalid_result: { type: integer }, missed_sla: { type: integer } } }
        dispute_window_secs: { type: integer }
        vault:
          type: object
          properties:
            address: { $ref: '#/components/schemas/Address' }
            deployed: { type: boolean }
            config_matches_manifest: { type: boolean }
            balance: { $ref: '#/components/schemas/Amount' }
            reserved: { $ref: '#/components/schemas/Amount' }
            runway_epochs: { type: integer }
            runway_days: { type: integer }
            total_paid: { $ref: '#/components/schemas/Amount' }
            pending_withdrawal: { $ref: '#/components/schemas/Amount' }
    ProjectStats:
      type: object
      properties:
        period: { type: string }
        miners_subscribed: { type: integer }
        miners_active: { type: integer }
        devices_by_class: { type: object, additionalProperties: { type: integer } }
        rounds: { type: integer }
        rounds_finalized: { type: integer }
        rounds_fallback: { type: integer }
        units: { type: integer }
        paid_to_miners: { $ref: '#/components/schemas/Amount' }
        avg_daily_reward_per_miner: { $ref: '#/components/schemas/Amount' }
        median_finality_ms: { type: integer }
        uptime_bp: { type: integer }
        slashes: { type: integer }
        growth_bp: { type: integer, description: 'Change in active miners vs previous period, signed bp' }
...